Index handler always sorts outputs [i.e. BaseProcessor.joint_sort() -now-> BaseIndex.joint_sort()]
